refactor(shared): extract denial verdict into a leaf module
documents.ts gained `import { isDenied } from "./operations.js"` for the
denied-operation branch in replayDocument. operations.ts already imported
nextSkipNumber/sortOperations back from documents.ts at runtime, so that
one value import closed a cycle and broke check-circular-imports. The
pre-existing documents.ts -> operations.ts edge was type-only, which dpdm
erases under -T, so nothing had closed the loop before.

The cycle was benign at runtime — every cross-file use sits inside a
function body, so no value is dereferenced during module evaluation — but
it fails the CI gate.

Move isDenied and the four *_REASON consensus constants into denied.ts.
They have no runtime dependencies, so the module cannot join a cycle: its
only import is a type-only Operation. Moving nextSkipNumber/sortOperations
into operations.ts instead would have been the wrong fix, since
nextSkipNumber calls garbageCollect and would re-open the cycle from the
other side.

The constants keep the grouping they were deliberately given, and no
consumer changes: they are all reached through the
@powerhousedao/shared/document-model barrel, which now re-exports
denied.js.

Verified: check-circular-imports clean (1134 files), tsc --build clean on
shared + reactor, reactor 2430/2430, document-model 343 passed.
